创建博客返回新博客的 id

+++ src/main/java/com/tinyblog/springboot/jdbctemplate/repository/BlogDao.java
@@ -16,8 +16,8 @@ public interface BlogDao {
     Blog findById(Long id);
     
     List<Blog> findByTitle(String title);
-    
-    int save(Blog blog);
+
+    Long save(Blog blog);
     
     int update(Blog blog);
 }
